refactored code, moved processing into ImportService and begun the process of calling packageIngestService
